Fixed Stuff!
Got things working so that we could save isHR and isManager to the cookie as well as username (username should probably be changed since grabbing it from the textbox is not ideal).

Also made the first to forms an off-white, it looks less jarring to me. What are your thoughts?