Issues on Salted Hashes
The queries for login are not working correctly. I don't know if i made a logicol error, or if my internet is just acting up and preventing the connection from working right.

I wanted to let someone else take a look. If you bypass the first query to get the hash, and uncomment //byte[] hashBytes = Convert.FromBase64String("4Nio0GMUEVdeRJ1vkS6Yog0rWUkqTZ/uOLGIds4KAodN1gMr");

it will execute to marker 6 and then fail on the next query it is asked to make and never get to marker 7. (Note that the markers will stop execution if they are hit since they change the username text and will mess up the next query in line)